change type

+++ phase1/src/TransferInable.java
@@ -1,4 +1,4 @@
 /** a transferin_able interface for the accounts which could be transferin */
 public interface TransferInable {
-   void transferIn(int amount);
+   void transferIn(double amount);
 }
