+++ phase1/src/TransferOutable.java
@@ -1,4 +1,4 @@
 /** the interface for the account which could transfer out to other accounts*/
 public interface TransferOutable {
-    void transferOut(int amount);
+    void transferOut(double amount);
 }
